allow base contracts and libraries to be specified via import object
A contract can be imported via an import object. For example, say you have
a file called a.sol:

	contract A {}
	library L {}

Now, another file can import this via an import object:

	import 'a.sol' as IMP;

	contract C is IMP.A {
		using IMP.L for *;
		constructor() IMP.A() {}
	}
