+++ tests/contract_testcases/solana/event.dot
@@ -3,15 +3,15 @@ strict digraph "tests/contract_testcases/solana/event.sol" {
 	contract [label="contract c\ntests/contract_testcases/solana/event.sol:4:1-5:12"]
 	foo [label="function foo\ncontract: c\ntests/contract_testcases/solana/event.sol:6:5-26\nsignature foo()\nvisibility public\nmutability nonpayable"]
 	diagnostic [label="found contract 'c'\nlevel Debug\ntests/contract_testcases/solana/event.sol:4:1-5:12"]
-	diagnostic_8 [label="'a' is not found\nlevel Error\ntests/contract_testcases/solana/event.sol:7:18-19"]
-	diagnostic_9 [label="'b' is not found\nlevel Error\ntests/contract_testcases/solana/event.sol:7:21-22"]
-	diagnostic_10 [label="'a' is not found\nlevel Error\ntests/contract_testcases/solana/event.sol:8:24-25"]
-	diagnostic_11 [label="'b' is not found\nlevel Error\ntests/contract_testcases/solana/event.sol:8:33-34"]
+	diagnostic_8 [label="'a' not found\nlevel Error\ntests/contract_testcases/solana/event.sol:7:18-19"]
+	diagnostic_9 [label="'b' not found\nlevel Error\ntests/contract_testcases/solana/event.sol:7:21-22"]
+	diagnostic_10 [label="'a' not found\nlevel Error\ntests/contract_testcases/solana/event.sol:8:24-25"]
+	diagnostic_11 [label="'b' not found\nlevel Error\ntests/contract_testcases/solana/event.sol:8:33-34"]
 	diagnostic_12 [label="duplicate argument with name 'no'\nlevel Error\ntests/contract_testcases/solana/event.sol:8:36-38"]
 	diagnostic_13 [label="'c' is a contract\nlevel Error\ntests/contract_testcases/solana/event.sol:8:40-41"]
 	diagnostic_14 [label="event 'Foo2' not found\nlevel Error\ntests/contract_testcases/solana/event.sol:9:14-18"]
-	diagnostic_15 [label="'a' is not found\nlevel Error\ntests/contract_testcases/solana/event.sol:9:25-26"]
-	diagnostic_16 [label="'b' is not found\nlevel Error\ntests/contract_testcases/solana/event.sol:9:34-35"]
+	diagnostic_15 [label="'a' not found\nlevel Error\ntests/contract_testcases/solana/event.sol:9:25-26"]
+	diagnostic_16 [label="'b' not found\nlevel Error\ntests/contract_testcases/solana/event.sol:9:34-35"]
 	diagnostic_17 [label="duplicate argument with name 'no'\nlevel Error\ntests/contract_testcases/solana/event.sol:9:37-39"]
 	diagnostic_18 [label="'c' is a contract\nlevel Error\ntests/contract_testcases/solana/event.sol:9:41-42"]
 	events -> Foo
