+++ tests/contract_testcases/solana/garbage_function_args.dot
@@ -10,11 +10,11 @@ strict digraph "tests/contract_testcases/solana/garbage_function_args.sol" {
 	g_10 [label="function g\ncontract: c\ntests/contract_testcases/solana/garbage_function_args.sol:14:2-25\nsignature g(bool)\nvisibility public\nmutability nonpayable"]
 	parameters_11 [label="parameters\nbool "]
 	diagnostic [label="found contract 'c'\nlevel Debug\ntests/contract_testcases/solana/garbage_function_args.sol:1:1-12"]
-	diagnostic_14 [label="'rubbish' is not found\nlevel Error\ntests/contract_testcases/solana/garbage_function_args.sol:3:11-18"]
-	diagnostic_15 [label="'meh' is not found\nlevel Error\ntests/contract_testcases/solana/garbage_function_args.sol:6:15-18"]
-	diagnostic_16 [label="'foo' is not found\nlevel Error\ntests/contract_testcases/solana/garbage_function_args.sol:9:9-12"]
-	diagnostic_17 [label="'oo' is not found\nlevel Error\ntests/contract_testcases/solana/garbage_function_args.sol:12:10-12"]
-	diagnostic_18 [label="'foo' is not found\nlevel Error\ntests/contract_testcases/solana/garbage_function_args.sol:15:14-17"]
+	diagnostic_14 [label="'rubbish' not found\nlevel Error\ntests/contract_testcases/solana/garbage_function_args.sol:3:11-18"]
+	diagnostic_15 [label="'meh' not found\nlevel Error\ntests/contract_testcases/solana/garbage_function_args.sol:6:15-18"]
+	diagnostic_16 [label="'foo' not found\nlevel Error\ntests/contract_testcases/solana/garbage_function_args.sol:9:9-12"]
+	diagnostic_17 [label="'oo' not found\nlevel Error\ntests/contract_testcases/solana/garbage_function_args.sol:12:10-12"]
+	diagnostic_18 [label="'foo' not found\nlevel Error\ntests/contract_testcases/solana/garbage_function_args.sol:15:14-17"]
 	contracts -> contract
 	contract -> g [label="function"]
 	g -> parameters [label="parameters"]
