+++ tests/contract_testcases/solana/import_contracts_via_object.dot
@@ -1,0 +1,22 @@
+strict digraph "tests/contract_testcases/solana/import_contracts_via_object.sol" {
+	contract [label="contract C\ntests/contract_testcases/solana/import_contracts_via_object.sol:2:1-3:20"]
+	base [label="base A\ntests/contract_testcases/solana/import_contracts_via_object.sol:3:15-20"]
+	using [label="using L"]
+	node_5 [label="constructor \ncontract: C\ntests/contract_testcases/solana/import_contracts_via_object.sol:5:2-23\nsignature ()\nvisibility public\nmutability nonpayable"]
+	contract_6 [label="contract A\ntests/contract_testcases/solana/simple.sol:1:1-12"]
+	contract_7 [label="contract L\ntests/contract_testcases/solana/simple.sol:1:14-2:11"]
+	diagnostic [label="found contract 'A'\nlevel Debug\ntests/contract_testcases/solana/simple.sol:1:1-12"]
+	diagnostic_10 [label="found library 'L'\nlevel Debug\ntests/contract_testcases/solana/simple.sol:1:14-2:11"]
+	diagnostic_11 [label="found contract 'C'\nlevel Debug\ntests/contract_testcases/solana/import_contracts_via_object.sol:2:1-3:20"]
+	diagnostic_12 [label="revert with custom error 'IMP.E' not supported yet\nlevel Error\ntests/contract_testcases/solana/import_contracts_via_object.sol:6:10-15"]
+	contracts -> contract
+	contract -> base [label="base"]
+	contract -> using [label="base"]
+	contract -> node_5 [label="constructor"]
+	contracts -> contract_6
+	contracts -> contract_7
+	diagnostics -> diagnostic [label="Debug"]
+	diagnostics -> diagnostic_10 [label="Debug"]
+	diagnostics -> diagnostic_11 [label="Debug"]
+	diagnostics -> diagnostic_12 [label="Error"]
+}
