+++ tests/contract_testcases/solana/import_contracts_via_object.sol
@@ -1,0 +1,8 @@
+import "simple.sol" as IMP;
+
+contract C is IMP.A {
+	using IMP.L for *;
+	constructor() IMP.A() {
+		revert IMP.E();
+	}
+}
