+++ tests/contract_testcases/solana/now.dot
@@ -16,10 +16,10 @@ strict digraph "tests/contract_testcases/solana/now.sol" {
 	return [label="return\ntests/contract_testcases/solana/now.sol:17:3-13"]
 	variable [label="variable: now\nint256\ntests/contract_testcases/solana/now.sol:17:10-13"]
 	diagnostic [label="found contract 'c'\nlevel Debug\ntests/contract_testcases/solana/now.sol:1:1-12"]
-	diagnostic_20 [label="'now' is not found\nlevel Error\ntests/contract_testcases/solana/now.sol:4:10-13"]
+	diagnostic_20 [label="'now' not found\nlevel Error\ntests/contract_testcases/solana/now.sol:4:10-13"]
 	diagnostic_21 [label="unknown function or type 'now'\nlevel Error\ntests/contract_testcases/solana/now.sol:7:3-6"]
-	diagnostic_22 [label="'now' is not found. 'now' was an alias for 'block.timestamp' in older versions of the Solidity language. Please use 'block.timestamp' instead.\nlevel Error\ntests/contract_testcases/solana/now.sol:10:10-13"]
-	diagnostic_23 [label="'now' is not found. 'now' was an alias for 'block.timestamp' in older versions of the Solidity language. Please use 'block.timestamp' instead.\nlevel Error\ntests/contract_testcases/solana/now.sol:13:10-13"]
+	diagnostic_22 [label="'now' not found. 'now' was an alias for 'block.timestamp' in older versions of the Solidity language. Please use 'block.timestamp' instead.\nlevel Error\ntests/contract_testcases/solana/now.sol:10:10-13"]
+	diagnostic_23 [label="'now' not found. 'now' was an alias for 'block.timestamp' in older versions of the Solidity language. Please use 'block.timestamp' instead.\nlevel Error\ntests/contract_testcases/solana/now.sol:13:10-13"]
 	structs -> S
 	contracts -> contract
 	contract -> f [label="function"]
