+++ tests/contract_testcases/solana/simple.sol
@@ -1,0 +1,2 @@
+contract A {}
+library L {}
