+++ tests/contract_testcases/substrate/expressions/destructure_01.dot
@@ -3,7 +3,7 @@ strict digraph "tests/contract_testcases/substrate/expressions/destructure_01.so
 	foo [label="function foo\ncontract: test\ntests/contract_testcases/substrate/expressions/destructure_01.sol:2:13-42\nsignature foo(uint256)\nvisibility public\nmutability nonpayable"]
 	parameters [label="parameters\nuint256 bar"]
 	diagnostic [label="found contract 'test'\nlevel Debug\ntests/contract_testcases/substrate/expressions/destructure_01.sol:1:1-15"]
-	diagnostic_6 [label="'c' is not found\nlevel Error\ntests/contract_testcases/substrate/expressions/destructure_01.sol:6:18-19"]
+	diagnostic_6 [label="'c' not found\nlevel Error\ntests/contract_testcases/substrate/expressions/destructure_01.sol:6:18-19"]
 	contracts -> contract
 	contract -> foo [label="function"]
 	foo -> parameters [label="parameters"]
