+++ tests/contract_testcases/substrate/functions/scopes_01.dot
@@ -3,7 +3,7 @@ strict digraph "tests/contract_testcases/substrate/functions/scopes_01.sol" {
 	goodset [label="function goodset\ncontract: test\ntests/contract_testcases/substrate/functions/scopes_01.sol:3:9-49\nsignature goodset()\nvisibility public\nmutability nonpayable"]
 	returns [label="returns\nuint256 "]
 	diagnostic [label="found contract 'test'\nlevel Debug\ntests/contract_testcases/substrate/functions/scopes_01.sol:2:5-19"]
-	diagnostic_6 [label="'i' is not found\nlevel Error\ntests/contract_testcases/substrate/functions/scopes_01.sol:7:20-21"]
+	diagnostic_6 [label="'i' not found\nlevel Error\ntests/contract_testcases/substrate/functions/scopes_01.sol:7:20-21"]
 	contracts -> contract
 	contract -> goodset [label="function"]
 	goodset -> returns [label="returns"]
