+++ tests/contract_testcases/substrate/inheritance/inherit_01.dot
@@ -1,6 +1,10 @@
 strict digraph "tests/contract_testcases/substrate/inheritance/inherit_01.sol" {
-	contract [label="contract a\ntests/contract_testcases/substrate/inheritance/inherit_01.sol:2:9-26"]
-	diagnostic [label="contract 'foo' not found\nlevel Error\ntests/contract_testcases/substrate/inheritance/inherit_01.sol:2:23-26"]
+	contract [label="contract a\ntests/contract_testcases/substrate/inheritance/inherit_01.sol:1:9-26"]
+	contract_3 [label="contract b\ntests/contract_testcases/substrate/inheritance/inherit_01.sol:6:9-30"]
+	diagnostic [label="'foo' not found\nlevel Error\ntests/contract_testcases/substrate/inheritance/inherit_01.sol:1:23-26"]
+	diagnostic_6 [label="'bar' not found\nlevel Error\ntests/contract_testcases/substrate/inheritance/inherit_01.sol:6:23-26"]
 	contracts -> contract
+	contracts -> contract_3
 	diagnostics -> diagnostic [label="Error"]
+	diagnostics -> diagnostic_6 [label="Error"]
 }
