+++ tests/contract_testcases/substrate/inheritance/inherit_01.sol
@@ -1,6 +1,9 @@
-
         contract a is foo {
             constructor(int arg1) public {
             }
         }
-        
+
+        contract b is bar.foo {
+            constructor(int arg1) public {
+            }
+        }
