+++ tests/contract_testcases/substrate/inheritance/inherit_variables_01.dot
@@ -7,7 +7,7 @@ strict digraph "tests/contract_testcases/substrate/inheritance/inherit_variables
 	returns [label="returns\nint256 "]
 	diagnostic [label="found contract 'b'\nlevel Debug\ntests/contract_testcases/substrate/inheritance/inherit_variables_01.sol:2:9-20"]
 	diagnostic_9 [label="found contract 'c'\nlevel Debug\ntests/contract_testcases/substrate/inheritance/inherit_variables_01.sol:6:9-24"]
-	diagnostic_10 [label="'foo' is not found\nlevel Error\ntests/contract_testcases/substrate/inheritance/inherit_variables_01.sol:8:24-27"]
+	diagnostic_10 [label="'foo' not found\nlevel Error\ntests/contract_testcases/substrate/inheritance/inherit_variables_01.sol:8:24-27"]
 	contracts -> contract
 	contract -> var [label="variable"]
 	contracts -> contract_3
